Updated Player class

+++ src/edu/gsu/csc1302/emperorsofspades/player/Player.java
@@ -74,7 +74,7 @@ public abstract class Player {
      * Logic is differed to the concrete player implementations.
      * @return a bid [4, 10].
      */
-    public abstract int placeBid();
+    public abstract double placeBid();
 
     /**
      * Return the number of cards in the user's deck of cards.
